+++ core/src/main/java/com/opensymphony/xwork2/factory/DefaultResultFactory.java
@@ -73,8 +73,8 @@ public class DefaultResultFactory implements ResultFactory {
 
             if (o instanceof Result) {
                 result = (Result) o;
-            } else if (o instanceof org.apache.struts2.Result) {
-                result = Result.adapt((org.apache.struts2.Result) o);
+            } else if (o instanceof org.apache.struts2.result.Result) {
+                result = Result.adapt((org.apache.struts2.result.Result) o);
             }
             if (result == null) {
                 throw new ConfigurationException("Class [" + resultClassName + "] does not implement Result", resultConfig);
