+++ core/src/main/java/org/apache/struts2/interceptor/ChainingInterceptor.java
@@ -27,9 +27,9 @@ import com.opensymphony.xwork2.util.reflection.ReflectionProvider;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.struts2.ActionInvocation;
-import org.apache.struts2.Result;
 import org.apache.struts2.StrutsConstants;
 import org.apache.struts2.Unchainable;
+import org.apache.struts2.result.Result;
 import org.apache.struts2.util.ValueStack;
 
 import java.util.ArrayList;
