+++ core/src/main/java/org/apache/struts2/result/Result.java
@@ -16,7 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2;
+package org.apache.struts2.result;
+
+import org.apache.struts2.Action;
+import org.apache.struts2.ActionInvocation;
 
 import java.io.Serializable;
