+++ core/src/test/java/com/opensymphony/xwork2/config/providers/ConfigurationProviderOgnlAllowlistTest.java
@@ -66,7 +66,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.SimpleAction"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
-                Class.forName("org.apache.struts2.Result"),
+                Class.forName("org.apache.struts2.result.Result"),
                 Class.forName("org.apache.struts2.Action"),
                 Class.forName("org.apache.struts2.Validateable"),
                 Class.forName("org.apache.struts2.interceptor.ValidationAware")
@@ -98,7 +98,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.SimpleAction"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
-                Class.forName("org.apache.struts2.Result"),
+                Class.forName("org.apache.struts2.result.Result"),
                 Class.forName("org.apache.struts2.Action"),
                 Class.forName("org.apache.struts2.Validateable"),
                 Class.forName("org.apache.struts2.interceptor.ValidationAware")
@@ -129,7 +129,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.Result"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
-                Class.forName("org.apache.struts2.Result"),
+                Class.forName("org.apache.struts2.result.Result"),
                 Class.forName("org.apache.struts2.Action"),
                 Class.forName("org.apache.struts2.Validateable"),
                 Class.forName("org.apache.struts2.interceptor.ValidationAware")
